pyproject.toml
[tool.poetry]
name = "kontext-copilot"
version = "0.1.0"
description = "Kontext Copilot is an AI empowered tool for data analytics that runs on your local computer."
authors = ["Kontext <enquiry@kontext.tech>"]
readme = "README.md"
homepage = "https://kontext.tech"
repository = "https://github.com/kontext-tech/kontext-ai"
license = "MIT"
# Include the static website folder in the package
include = ["kontext_copilot/ui/**/*", "kontext_copilot/api/**/*.json"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.111.0"
python-dotenv = "^1.0.1"
pre-commit = "^3.7.1"
typing-extensions = ">=4.10.0,<5.0.0"
sqlalchemy = "^2.0.31"
alembic = "^1.13.2"

pydantic = "^2.8.2"
[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "$version"
version_scheme = "semver2"
version_provider = "poetry"
update_changelog_on_bump = true
major_version_zero = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = ">=7.4.3,<9.0.0"
pytest-asyncio = "^0.23.2"
pytest-cov = ">=4.1,<6.0"
pytest-httpserver = "^1.0.8"
pillow = "^10.2.0"
ruff = ">=0.1.8,<0.5.0"

commitizen = "^3.27.0"
[tool.ruff]
line-length = 999
indent-width = 2

[tool.ruff.format]
quote-style = "single"
indent-style = "space"

[tool.ruff.lint]
select = ["E", "F", "B"]
ignore = ["E501"]

[tool.pytest.ini_options]
addopts = '--doctest-modules --ignore examples'
